Set BoardModerator to BoardAdmin (with the few minor changes that needed to be made)

+++ Toucan/client/src/components/BoardModerator.vue
@@ -1,48 +1,63 @@
 <template>
-<div id="bg">
-  <div class="container">
-    <NavbarComponent />
-    <div v-if="isMod" class="row">
-      <div class="col-lg-4">
-        <AddCourse @add-course="addCourse"/>
-        <Courses
-        @delete-course="deleteCourse"
-        @show-posts="showPosts"
-        @show-all-posts="showAllPosts"
-        :courses="courses"
-        :canEdit="canEdit"
-        />
+  <div id="bg">
+    <div class="container">
+      <NavbarComponent />
+      <div v-if="isMod" class="row">
+        <div class="col-lg-4">
+          <AddCourse @add-course="addCourse" />
+          <Courses
+            @delete-course="deleteCourse"
+            @show-posts="showPosts"
+            @show-all-posts="showAllPosts"
+            :courses="courses"
+            :canEdit="canEdit"
+          />
+        </div>
+        <div class="col-lg-8">
+          <div id="post-header"></div>
+          <CourseTabs
+            @load-more-posts="
+              numShownPosts += 5;
+              filteredPosts = showNumPosts(isSelected ? coursePosts : posts);
+            "
+            :posts="filteredPosts"
+            @delete-post="deletePost"
+            @re-render-posts="forceRerender"
+            :name="selectedName"
+            :course="selectedCourse"
+            :isSelected="isSelected"
+            :hasManyPosts="hasManyPosts"
+            :canEdit="canEdit"
+            :members="courseMembers"
+            :details="courseDetails"
+          />
+        </div>
       </div>
-      <div class="col-lg-8">
-        <div id="post-header"></div>
-        <CourseTabs :posts="filteredPosts" @delete-post="deletePost" @re-render-posts="forceRerender" :course="selectedCourse" :isSelected="isSelected"  :hasManyPosts="hasManyPosts" :canEdit="canEdit" :members="courseMembers" :details="courseDetails"/>
-      </div>  
-    </div>
-    <div class="row" v-if="!isMod">
-      <h3>{{ content }}</h3>
+      <div class="row" v-if="!isMod">
+        <h3>{{ content }}</h3>
+      </div>
     </div>
   </div>
-</div>
 </template>
 
 <script>
-import AddCourse from '@/components/AddCourse.vue'
-import Courses from '@/components/Courses.vue'
-import NavbarComponent from '@/components/NavbarComponent.vue'
-import CourseTabs from '@/components/CourseTabs.vue'
+import AddCourse from "@/components/AddCourse.vue";
+import Courses from "@/components/Courses.vue";
+import NavbarComponent from "@/components/NavbarComponent.vue";
+import CourseTabs from "@/components/CourseTabs.vue";
 
-import PostService from '@/PostService.js'
-import CourseService from '@/CourseService.js'
+import PostService from "@/PostService.js";
+import CourseService from "@/CourseService.js";
 
-import UserService from "../services/user.service"
+import UserService from "../services/user.service";
 
 export default {
-  name: 'Home',
+  name: "Home",
   components: {
     AddCourse,
     Courses,
     NavbarComponent,
-    CourseTabs
+    CourseTabs,
   },
   data() {
     return {
@@ -50,19 +65,29 @@ export default {
       posts: [],
       coursePosts: [],
       filteredPosts: [],
-      selectedCourse: '',
+      selectedCourse: "",
+      selectedName: "",
       isSelected: false,
       hasManyPosts: false,
       content: "",
       isMod: false,
       canEdit: true,
+      numShownPosts: 5,
       courseMembers: [],
       courseDetails: [],
     }
   },
   methods: {
     async addCourse(course) {
-      await CourseService.insertCourse(course.name, course.courseCode, course.teacher, course.members, course.time, course.img, course.modalID);
+      await CourseService.insertCourse(
+        course.name,
+        course.courseCode,
+        course.teacher,
+        course.members,
+        course.time,
+        course.img,
+        course.modalID
+      );
       this.courses = await CourseService.getCourses();
     },
     async deleteCourse(id) {
@@ -70,13 +95,12 @@ export default {
       this.courses = await CourseService.getCourses(); // Gets all the courses again now that a specific course has been deleted
 
       // Find all the posts that belong to the deleted course and deletes them one by one
-      this.posts.forEach(async post => {
-        if (post.courseID === id)
-          await PostService.deletePost(post._id);
+      this.posts.forEach(async (post) => {
+        if (post.courseID === id) await PostService.deletePost(post._id);
       });
       
       this.posts = (await PostService.getPosts()).reverse();  // Gets all the posts again now that all the posts belonging to the deleted course have been deleted
-      this.filteredPosts = this.showFivePosts(this.posts); // Sets filteredPosts (determines which posts are displayed) to all remaining posts
+      this.filteredPosts = this.showNumPosts(this.posts); // Sets filteredPosts (determines which posts are displayed) to all remaining posts
     },
     async showPosts(id) {
       // old version of showPosts without using component; may revert back to this if posts component fails
@@ -98,69 +122,74 @@ export default {
           document.querySelector('#posts').appendChild(coursePost);
         }
       })*/
+      this.selectedName = this.getCourseName(id);
       this.selectedCourse = id;
       this.isSelected = true;
+      this.numShownPosts = 5;
       this.findCourseMembers();
       this.findCourseDetails();
-      this.coursePosts = this.posts.filter(post => (post.courseID === id));
-      this.filteredPosts = this.showFivePosts(this.coursePosts);
+      this.coursePosts = this.posts.filter((post) => post.courseID === id);
+      this.filteredPosts = this.showNumPosts(this.coursePosts);
     },
     showAllPosts() {
       this.isSelected = false;
-      this.filteredPosts = this.showFivePosts(this.posts);
+      this.numShownPosts = 5;
+      this.filteredPosts = this.showNumPosts(this.posts);
     },
     getCourseName(id) {
-      let courseName = '';
+      let courseName = "";
       this.courses.forEach((course) => {
         if (course._id === id) {
           courseName = course.name;
         }
-      })
+      });
       return courseName;
     },
     async deletePost(id) {
       console.log(id);
 
       let courseOfPost;
-      this.posts.forEach(post => {
-        if (post._id === id)
-          courseOfPost = post.courseID;
+      this.posts.forEach((post) => {
+        if (post._id === id) courseOfPost = post.courseID;
       });
 
       await PostService.deletePost(id);
 
       this.posts = (await PostService.getPosts()).reverse();
 
-      if (this.isSelected) // Makes it so that if a course is selected, displays those posts, otherwise dispaly all posts
-        this.filteredPosts = this.posts.filter(post => (post.courseID === courseOfPost));
-      else
-        this.filteredPosts = this.posts;
+      if (this.isSelected)
+        // Makes it so that if a course is selected, displays those posts, otherwise dispaly all posts
+        this.filteredPosts = this.posts.filter(
+          (post) => post.courseID === courseOfPost
+        );
+      else this.filteredPosts = this.posts;
     },
     async forceRerender(id) {
       this.posts = (await PostService.getPosts()).reverse();
-      this.showPosts(id)
+      this.showPosts(id);
     },
-    showFivePosts(posts) {
+    showNumPosts(posts) {
       this.hasManyPosts = false;
 
       let postsToShow = [];
 
       posts.forEach((post, index) => {
-        if (index < 5)
+        if (index < this.numShownPosts)
           postsToShow.push(JSON.parse(JSON.stringify(post)));
       });
 
-      if (posts.length > 5) {
+      if (posts.length > this.numShownPosts) {
         this.hasManyPosts = true;
       }
 
       return postsToShow;
     },
     findCourseMembers() {
+      this.courseMembers = [];
       this.courses.forEach((course) => {
         if (course._id === this.selectedCourse)
           this.courseMembers = course.members.split(";");
-      })
+      });
     },
     findCourseDetails() {
       this.courseDetails = [];
@@ -170,20 +199,20 @@ export default {
           this.courseDetails.push(`Code: ${course.courseCode}`);
           this.courseDetails.push(`Teacher: ${course.teacher}`);
         }
-      })
+      });
       this.courseDetails = JSON.parse(JSON.stringify(this.courseDetails));
-    }
+    },
   },
   async created() {
     this.courses = await CourseService.getCourses();
     this.posts = (await PostService.getPosts()).reverse();
-    this.filteredPosts = this.showFivePosts(this.posts);
+    this.filteredPosts = this.showNumPosts(this.posts);
   },
   mounted() {
-    UserService.getAdminBoard().then(
+    UserService.getModeratorBoard().then(
       (response) => {
         this.isMod = true;
-        console.log('test');
+        console.log("test");
       },
       (error) => {
         this.content =
@@ -195,7 +224,7 @@ export default {
       }
     );
   },
-}
+};
 </script>
 
 <style scoped>
@@ -212,12 +241,11 @@ div.post-title {
 </style>
 
 <style>
-
-body{
-    background: url(../assets/homebg.png) no-repeat;
-    background-size: cover;
-    background-position: center;
-    background-attachment: fixed;
-    height: 100vh;
-  }
+body {
+  background: url(../assets/homebg.png) no-repeat;
+  background-size: cover;
+  background-position: center;
+  background-attachment: fixed;
+  height: 100vh;
+}
 </style>
